First Steps
This already selects, opens, gets and stores the text from all the links
(in theory, no real tests have been done).

Currently it only does it with articles that are in Yahoo webpage. (It
click read more if you're curious).

Next step is to make which-website-is-opened algoryth to select HOW to
get the data.

Also it would be great to make these files (article text) as easy as
possible to find, open and use for anything you want, but that's in the
future, maybe with a different project.