New approach
I've decided to scrape UIPath approach as it was REALLY limited to what
i've needed.

Currently this gets redirected urls and raw text from the articles. It
seems to work just fine.

Next step is to update the results file in order to make an easier
matching when we need the actual stock price numbers.

Also clean up of the text, saving files etc. are in due order.